src/main/java/guru/springframework/msscbeerservice/web/controller/BeerController.java
package guru.springframework.msscbeerservice.web.controller;

import guru.springframework.msscbeerservice.web.model.BeerDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

@RequestMapping("api/v1/beer")
@RestController
public class BeerController {

    @GetMapping({"/{beerId}"})
    public ResponseEntity getBeerById(@PathVariable("beerId") UUID beerId){
        // TODO: 04-07-2020  
        return new ResponseEntity(BeerDto.builder().build(), HttpStatus.OK);
    }

    @PostMapping
    public ResponseEntity<BeerDto> saveNewBeer(@RequestBody BeerDto beerDto){
        // TODO: 04-07-2020  
        return new ResponseEntity<>(HttpStatus.CREATED);
    }

    @PutMapping({"/{beerId}"})
    public ResponseEntity<BeerDto> updateBeer(@RequestBody BeerDto beerDto, @PathVariable("beerId") UUID beerId){
        // TODO: 04-07-2020
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    
}
